Move blacklight_advanced_search generator to blacklight_advanced_search:install
